Refactor BaseMViewTemplate and MysqlMViewTemplate
