+++ libs/db-browser/src/main/java/com/oceanbase/tools/dbbrowser/template/mysql/MysqlMViewTemplate.java
@@ -23,6 +23,7 @@ import java.util.stream.Collectors;
 
 import javax.validation.constraints.NotNull;
 
+import com.oceanbase.tools.dbbrowser.template.BaseMViewTemplate;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.Validate;
@@ -46,104 +47,33 @@ import com.oceanbase.tools.dbbrowser.util.SqlBuilder;
  * @date: 2025/3/10 16:45
  * @since: 4.3.4
  */
-public class MysqlMViewTemplate implements DBObjectTemplate<DBMaterializedView> {
-    private MySQLViewTemplate mySQLViewTemplate;
-
-    private DBTableConstraintEditor dbTableConstraintEditor;
-
-    private DBTablePartitionEditor dbTablePartitionEditor;
+public class MysqlMViewTemplate extends BaseMViewTemplate {
 
     public MysqlMViewTemplate() {
-        mySQLViewTemplate = new MySQLViewTemplate();
-        dbTableConstraintEditor = new MySQLConstraintEditor();
-        dbTablePartitionEditor = new OBMySQLDBTablePartitionEditor();
+        super(new MySQLViewTemplate(),new MySQLConstraintEditor(),new OBMySQLDBTablePartitionEditor());
     }
 
     @Override
-    public String generateCreateObjectTemplate(DBMaterializedView dbObject) {
-        Validate.notBlank(dbObject.getName(), "Materialized view name can not be blank");
-        DBView dbView = dbObject.generateDBView();
-        mySQLViewTemplate.validOperations(dbView);
-        SqlBuilder sqlBuilder = new MySQLSqlBuilder();
-        sqlBuilder.append("create materialized view ")
-                .append(getFullyQualifiedTableName(dbObject));
-        // build sql about primary key
-        if (CollectionUtils.isNotEmpty(dbObject.getConstraints())) {
-            Validate.isTrue(
-                    dbObject.getConstraints().size() == 1 && dbObject.getConstraints().get(0).getType() == PRIMARY_KEY,
-                    "Only primary key is supported");
-            DBTableConstraint dbTableConstraint = dbObject.getConstraints().get(0);
-            sqlBuilder.append("(")
-                    .append(getPrimary(dbTableConstraintEditor.generateCreateDefinitionDDL(dbTableConstraint)))
-                    .append(")");
-        }
-        // build sql about parallelism degree
-        if (Objects.nonNull(dbObject.getParallelismDegree()) && dbObject.getParallelismDegree() > 1) {
-            sqlBuilder.line().append("PARALLEL ").append(dbObject.getParallelismDegree());
-        }
-        // build sql about partition
-        if (Objects.nonNull(dbObject.getPartition())) {
-            sqlBuilder.line()
-                    .append(dbTablePartitionEditor.generateCreateDefinitionDDL(dbObject.getPartition()));
-        }
-        // build sql about column group
-        if (CollectionUtils.isNotEmpty(dbObject.getColumnGroups())) {
-            sqlBuilder.line().append(" WITH COLUMN GROUP(")
-                    .append(dbObject.getColumnGroups().stream().map(DBColumnGroupElement::toString)
-                            .collect(Collectors.joining(",")))
-                    .append(")");
-        }
-        // build sql about refresh method
-        if (Objects.nonNull(dbObject.getRefreshMethod())) {
-            sqlBuilder.line().append(dbObject.getRefreshMethod().getCreateName());
-        }
-        // build sql about refresh schedule
-        if (Objects.nonNull(dbObject.getRefreshSchedule())) {
-            DBMaterializedViewRefreshSchedule refreshSchedule = dbObject.getRefreshSchedule();
+    protected SqlBuilder sqlBuilder() {
+        return new MySQLSqlBuilder();
+    }
+
+    @Override
+    protected void fillRefreshSchedule(DBMaterializedView dbMView, SqlBuilder sqlBuilder) {
+        if (Objects.nonNull(dbMView.getRefreshSchedule())) {
+            DBMaterializedViewRefreshSchedule refreshSchedule = dbMView.getRefreshSchedule();
             if (refreshSchedule.getStartStrategy() == DBMaterializedViewRefreshSchedule.StartStrategy.START_NOW) {
                 sqlBuilder.line().append("START WITH sysdate()");
                 sqlBuilder.line().append("NEXT sysdate() + INTERVAL ").append(refreshSchedule.getInterval()).append(" ")
-                        .append(refreshSchedule.getUnit());
+                    .append(refreshSchedule.getUnit());
             } else if (refreshSchedule.getStartStrategy() == DBMaterializedViewRefreshSchedule.StartStrategy.START_AT) {
                 SimpleDateFormat formatter = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
                 String formattedDate = formatter.format(refreshSchedule.getStartWith());
                 sqlBuilder.line().append("START WITH TIMESTAMP '").append(formattedDate).append("'");
                 sqlBuilder.line().append("NEXT TIMESTAMP '").append(formattedDate).append("' + INTERVAL ")
-                        .append(refreshSchedule.getInterval()).append(" ").append(refreshSchedule.getUnit());
+                    .append(refreshSchedule.getInterval()).append(" ").append(refreshSchedule.getUnit());
             }
         }
-        // build sql about query rewrite
-        if (Objects.nonNull(dbObject.getEnableQueryRewrite())) {
-            if (dbObject.getEnableQueryRewrite()) {
-                sqlBuilder.line().append("ENABLE QUERY REWRITE");
-            } else {
-                sqlBuilder.line().append("DISABLE QUERY REWRITE");
-            }
-        }
-        // build sql about query computation
-        if (Objects.nonNull(dbObject.getEnableQueryComputation())) {
-            if (dbObject.getEnableQueryComputation()) {
-                sqlBuilder.line().append("ENABLE ON QUERY COMPUTATION");
-            } else {
-                sqlBuilder.line().append("DISABLE ON QUERY COMPUTATION");
-            }
-        }
-        sqlBuilder.line().append("AS");
-        // build sql about query statement
-        mySQLViewTemplate.generateQueryStatement(dbView, sqlBuilder);
-        return sqlBuilder.toString();
     }
 
-    private String getPrimary(@NotNull String input) {
-        return input.replaceFirst("(?i)CONSTRAINT\\s*", "");
-    }
-
-    private String getFullyQualifiedTableName(@NotNull DBMaterializedView DBMaterializedView) {
-        SqlBuilder sqlBuilder = new MySQLSqlBuilder();;
-        if (StringUtils.isNotEmpty(DBMaterializedView.getSchemaName())) {
-            sqlBuilder.identifier(DBMaterializedView.getSchemaName()).append(".");
-        }
-        sqlBuilder.identifier(DBMaterializedView.getName());
-        return sqlBuilder.toString();
-    }
 }
