+++ libs/db-browser/src/test/java/com/oceanbase/tools/dbbrowser/template/MysqlMViewTemplateTest.java
@@ -64,25 +64,25 @@ public class MysqlMViewTemplateTest {
         DBMaterializedView.setOperations(Collections.singletonList("left join"));
         DBMaterializedView.setCreateColumns(prepareQueryColumns(2));
 
-        String expect = "create materialized view `schema_0`.`mv_0`(PRIMARY KEY (`alias_c0`))\n" +
-                "PARALLEL 8\n" +
-                " PARTITION BY HASH(`alias_c0`) \n" +
-                "PARTITIONS 3\n" +
-                " WITH COLUMN GROUP(all columns,each column)\n" +
-                "REFRESH COMPLETE\n" +
-                "START WITH sysdate()\n" +
-                "NEXT sysdate() + INTERVAL 1 DAY\n" +
-                "DISABLE QUERY REWRITE\n" +
-                "DISABLE ON QUERY COMPUTATION\n" +
-                "AS\n" +
-                "select\n" +
-                "\ttableAlias_0.`c_0` as alias_c0,\n" +
-                "\ttableAlias_0.`d_0` as alias_d0,\n" +
-                "\ttableAlias_1.`c_1` as alias_c1,\n" +
-                "\ttableAlias_1.`d_1` as alias_d1\n" +
-                "from\n" +
-                "\t`database_0`.`table_0` tableAlias_0\n" +
-                "\tleft join `database_1`.`table_1` tableAlias_1 on /* TODO enter attribute to join on here */";
+        String expect = "CREATE MATERIALIZED VIEW `schema_0`.`mv_0`(PRIMARY KEY (`alias_c0`))\n" +
+            "PARALLEL 8\n" +
+            " PARTITION BY HASH(`alias_c0`) \n" +
+            "PARTITIONS 3\n" +
+            " WITH COLUMN GROUP(all columns,each column)\n" +
+            "REFRESH COMPLETE\n" +
+            "START WITH sysdate()\n" +
+            "NEXT sysdate() + INTERVAL 1 DAY\n" +
+            "DISABLE QUERY REWRITE\n" +
+            "DISABLE ON QUERY COMPUTATION\n" +
+            "AS\n" +
+            "select\n" +
+            "\ttableAlias_0.`c_0` as alias_c0,\n" +
+            "\ttableAlias_0.`d_0` as alias_d0,\n" +
+            "\ttableAlias_1.`c_1` as alias_c1,\n" +
+            "\ttableAlias_1.`d_1` as alias_d1\n" +
+            "from\n" +
+            "\t`database_0`.`table_0` tableAlias_0\n" +
+            "\tleft join `database_1`.`table_1` tableAlias_1 on /* TODO enter attribute to join on here */";
         String actual = mysqlMViewTemplate.generateCreateObjectTemplate(DBMaterializedView);
         Assert.assertEquals(expect, actual);
     }
@@ -100,18 +100,19 @@ public class MysqlMViewTemplateTest {
         DBMaterializedView.setOperations(Collections.singletonList("left join"));
         DBMaterializedView.setCreateColumns(prepareQueryColumns(2));
 
-        String expect = "create materialized view `schema_0`.`mv_0`\n" +
-                "START WITH TIMESTAMP '2025-07-11 18:00:00'\n" +
-                "NEXT TIMESTAMP '2025-07-11 18:00:00' + INTERVAL 1 WEEK\n" +
-                "AS\n" +
-                "select\n" +
-                "\ttableAlias_0.`c_0` as alias_c0,\n" +
-                "\ttableAlias_0.`d_0` as alias_d0,\n" +
-                "\ttableAlias_1.`c_1` as alias_c1,\n" +
-                "\ttableAlias_1.`d_1` as alias_d1\n" +
-                "from\n" +
-                "\t`database_0`.`table_0` tableAlias_0\n" +
-                "\tleft join `database_1`.`table_1` tableAlias_1 on /* TODO enter attribute to join on here */";
+        String expect = "CREATE MATERIALIZED VIEW `schema_0`.`mv_0`\n" +
+            "REFRESH FORCE\n" +
+            "START WITH TIMESTAMP '2025-07-11 18:00:00'\n" +
+            "NEXT TIMESTAMP '2025-07-11 18:00:00' + INTERVAL 1 WEEK\n" +
+            "AS\n" +
+            "select\n" +
+            "\ttableAlias_0.`c_0` as alias_c0,\n" +
+            "\ttableAlias_0.`d_0` as alias_d0,\n" +
+            "\ttableAlias_1.`c_1` as alias_c1,\n" +
+            "\ttableAlias_1.`d_1` as alias_d1\n" +
+            "from\n" +
+            "\t`database_0`.`table_0` tableAlias_0\n" +
+            "\tleft join `database_1`.`table_1` tableAlias_1 on /* TODO enter attribute to join on here */";
         String actual = mysqlMViewTemplate.generateCreateObjectTemplate(DBMaterializedView);
         Assert.assertEquals(expect, actual);
     }
